backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Werkzeug==3.0.1
pandas==2.2.3
openpyxl==3.1.2
python-dotenv==1.0.0
bcrypt==4.1.2
requests==2.31.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
pytest==8.3.4
pytest-flask==1.3.0

